requirements.txt
paho-mqtt==1.6.1
openai==1.3.0